Add decryptor and AddFrame to Storage

+++ src/models.go
@@ -1,17 +1,17 @@
 package src
 
 type Frame struct {
-	Id                  int
+	Id                  int64
 	Time                string
-	Longitude           float32
-	Latitude            float32
+	Longitude           float64
+	Latitude            float64
 	Status              string
 	LatitudeHemisphere  string
 	LongitudeHemisphere string
-	EarthVelocity       float32
-	Track               float32
+	EarthVelocity       float64
+	Track               float64
 	Date                string
-	MagneticVariation   float32
+	MagneticVariation   float64
 	DirectionVariation  string
 	SystemPosition      string
 	Checksum            string
